Rydder opp i mappestrukur
